Trying out a new system 6/11/20
NEW METHOD OF CALCULATING BYE AFFINITY EACH WEEK

3 is now the "default" affinity.

If you receive a round 1 bye, your affinity is set to 1.
If you receive a round 2 bye, your affinity is set to 2.
If you receive a round 3 bye, your affinity is set to 3.

If your affinity was less than 3 and you did not get a bye, increment it by 1 (regardless of whether you were present).

[Effectively, getting a bye gives you a 3-week "affinity recharge" period, or shorter if it wasn't the round 1 bye. During the recharge period, your attendance will not affect your affinity. This will protect players who recently got the bye from getting it again for a while.]

If your affinity is 3+ and you were present and you did not get a bye, increment it by 1 for each round you played. (Usually 3, but if you drop and leave before the end of the draft for reasons unrelated to byes, it could be less than 3.)

[This is basically the same as the current system. The more you get to play without getting a bye, the more likely you are to get the bye.]

If your affinity is 3+ and you were not present, decrement it by 2^N-1 where N is the number of weeks in a row that you missed, down to a minimum of 3. To track this, whenever I decrement someone's affinity, I will add an indicator next to their name noting how many affinity points they would lose the following week if they don't show up, e.g. "Jon -2" meaning that Jon was absent this week and has lost 1 affinity point, and will lose 2 next week; then next week if Jon is absent again I would put "Jon -4" etc.

[Under the old system, affinity drained very slowly, even if you missed many weeks in a row. Under the new system, affinity will drain at an accelerated rate the more consecutive weeks you miss. This way, a player who shows up only rarely or who misses a lot of weeks will be protected from getting the bye their first time back.]